Add retention policy config for history.
It turns out that having a large tag key cardinality in influxdb is a really bad idea. Especially if you are trying to run on an underpowered machine like a raspberry pi. So having a configurable retention policy is absolutely necessary in case you want longer retention on the rest of the data in the database.